+++ src/SQLProvider.Runtime/Providers.Firebird.fs
@@ -765,6 +765,7 @@ type internal FirebirdProvider(resolutionPath, contextSchemaPath, owner, referen
                     | IndexOfStart(SqlCol(al2, col2),(SqlConstant startPos)) -> sprintf "POSITION(%s,%s,%s)" (fieldNotation al2 col2) column (fieldParam startPos)
                     | IndexOfStart(SqlCol(al2, col2),SqlCol(al3, col3)) -> sprintf "POSITION(%s,%s,%s)" (fieldNotation al2 col2) column (fieldNotation al3 col3)
                     | CastVarchar -> sprintf "CAST(%s AS CHAR)" column
+                    | CastInt -> sprintf "CAST(%s AS INT)" column
                     // Date functions
                     | Date -> sprintf "CAST (%s AS DATE)" column
                     | Year -> sprintf "EXTRACT(YEAR FROM %s)" column